Move the spotify reader test to integration tests
